+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fixed Well Pulley item model
 - Added Tool for Crafting Workspace recipe
 - Render fluid instead of filled bucket
+- Improved JEI/EMI Crafting Workspace recipe rendering
 
 ## [0.9.2]